moves and splits configs, removes form complete from pipelines
The hcvtarget and prioritize projects are more different than previously
thought. This commit splits and names their configs for clarity.

form complete fields are not able to be used with the flat json format
when there are missing required fields. This commit removes those steps
from the pipeline till a fix is found.